requirements.txt
psutil~=5.9.8
python-dotenv~=1.0.1
mysql-connector-python~=8.4.0